DERBY-4614: Ported 1042675 from trunk to 10.7 branch.
git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.7@1042682 13f79535-47bb-0310-9956-ffa450edef68

+++ java/testing/org/apache/derbyTesting/functionTests/tests/lang/TableFunctionTest.java
@@ -788,9 +788,9 @@ public class TableFunctionTest extends BaseJDBCTestCase
             FUNCTION_RESULT_COLUMN,
             JDBC_TYPE_TIMESTAMP,
             "TIMESTAMP",
-            new Integer( 26 ),     // PRECISION
+            new Integer( 29 ),     // PRECISION
             new Integer( 16 ),         // LENGTH
-            new Integer( 6 ),       // SCALE
+            new Integer( 9 ),       // SCALE
             RADIX_INTEGER,    // RADIX
             ALLOWS_NULLS,
             EMPTY_REMARKS,
